+++ src/main/java/org/gradlex/jvm/dependency/conflict/resolution/rules/AddApiDependencyMetadataRule.java
@@ -19,13 +19,17 @@ package org.gradlex.jvm.dependency.conflict.resolution.rules;
 import org.gradle.api.artifacts.CacheableRule;
 import org.gradle.api.artifacts.ComponentMetadataContext;
 import org.gradle.api.artifacts.ComponentMetadataRule;
+import org.gradle.api.attributes.Category;
+import org.gradle.api.attributes.Usage;
 
 import javax.inject.Inject;
+
+import static org.gradlex.jvm.dependency.conflict.resolution.rules.VariantSelection.allVariantsMatching;
 
 /**
  * See:
  * <a href="https://docs.gradle.org/current/userguide/component_metadata_rules.html#fixing_wrong_dependency_details">
- *     component_metadata_rules.html#fixing_wrong_dependency_details</a>
+ * component_metadata_rules.html#fixing_wrong_dependency_details</a>
  */
 @CacheableRule
 public abstract class AddApiDependencyMetadataRule implements ComponentMetadataRule {
@@ -39,6 +43,8 @@ public abstract class AddApiDependencyMetadataRule implements ComponentMetadataR
 
     @Override
     public void execute(ComponentMetadataContext context) {
-        context.getDetails().allVariants(v -> v.withDependencies(d -> d.add(dependency)));
+        allVariantsMatching(context,
+                id -> (id.matches(Usage.USAGE_ATTRIBUTE, Usage.JAVA_API) || id.matches(Usage.USAGE_ATTRIBUTE, Usage.JAVA_RUNTIME)) && id.matches(Category.CATEGORY_ATTRIBUTE, Category.LIBRARY),
+                v -> v.withDependencies(d -> d.add(dependency)));
     }
 }