+++ src/main/java/org/gradlex/jvm/dependency/conflict/resolution/rules/ReduceToCompileOnlyApiDependencyMetadataRule.java
@@ -19,14 +19,18 @@ package org.gradlex.jvm.dependency.conflict.resolution.rules;
 import org.gradle.api.artifacts.CacheableRule;
 import org.gradle.api.artifacts.ComponentMetadataContext;
 import org.gradle.api.artifacts.ComponentMetadataRule;
+import org.gradle.api.attributes.Category;
+import org.gradle.api.attributes.Usage;
 
 import javax.inject.Inject;
 import java.util.stream.Collectors;
 
+import static org.gradlex.jvm.dependency.conflict.resolution.rules.VariantSelection.allVariantsMatching;
+
 /**
  * See:
  * <a href="https://docs.gradle.org/current/userguide/component_metadata_rules.html#fixing_wrong_dependency_details">
- *     component_metadata_rules.html#fixing_wrong_dependency_details</a>
+ * component_metadata_rules.html#fixing_wrong_dependency_details</a>
  */
 @CacheableRule
 public abstract class ReduceToCompileOnlyApiDependencyMetadataRule implements ComponentMetadataRule {
@@ -40,7 +44,8 @@ public abstract class ReduceToCompileOnlyApiDependencyMetadataRule implements Co
 
     @Override
     public void execute(ComponentMetadataContext context) {
-        context.getDetails().withVariant("runtime", v -> v.withDependencies(d -> d.removeAll(d.stream().filter(it -> dependency.equals(it.getModule().toString())).collect(Collectors.toList())))); // .pom
-        context.getDetails().withVariant("runtimeElements", v -> v.withDependencies(d -> d.removeAll(d.stream().filter(it -> dependency.equals(it.getModule().toString())).collect(Collectors.toList())))); // .module
+        allVariantsMatching(context,
+                id -> id.matches(Usage.USAGE_ATTRIBUTE, Usage.JAVA_RUNTIME) && id.matches(Category.CATEGORY_ATTRIBUTE, Category.LIBRARY),
+                v -> v.withDependencies(d -> d.removeAll(d.stream().filter(it -> dependency.equals(it.getModule().toString())).collect(Collectors.toList()))));
     }
 }